rename blocks
