+++ blocks/gmo-program-header/gmo-program-header.css
@@ -6,7 +6,6 @@
 .inputs-wrapper {
     display: flex;
     height: 50px;
-    padding-left: 10px;
 }
 .search-wrapper {
     background-color: #FFF;
